+++ src/main/java/stringcalculator/Calculator.java
@@ -2,6 +2,8 @@ package stringcalculator;
 
 import java.util.Scanner;
 
+import stringcalculator.domain.StringArithmeticInput;
+
 public class Calculator {
     public static void main(String[] args) {
         System.out.println("계산식을 입력해주세요: ");
