+++ src/main/java/stringcalculator/command/CommandFactory.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package stringcalculator.command;
 
 import java.util.Set;
