+++ src/main/java/stringcalculator/domain/StringArithmeticInput.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package stringcalculator.domain;
 
 public class StringArithmeticInput {
     private static final String DELIMITER = " ";
