+++ src/test/java/stringcalculator/command/CommandFactoryTest.java
@@ -1,10 +1,11 @@
-package stringcalculator;
+package stringcalculator.command;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import stringcalculator.command.AddCommand;
+import stringcalculator.command.CommandFactory;
 
 public class CommandFactoryTest {
