+++ src/test/java/stringcalculator/domain/StringArithmeticInputTestTest.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package stringcalculator.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -6,6 +6,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+
+import stringcalculator.domain.Number;
+import stringcalculator.domain.StringArithmeticInput;
 
 public class StringArithmeticInputTestTest {
